This is a better way of defining args. Trust me...
if config.prefix is `ok google,` the original code would have split `ok` as the command and prefix, and tried to slice off 10 chars off that as the command and prefix combo. Additionally, it allows if you were to define the prefix as `<@123456789101112>` that both that mention and mention + space would work as a prefix. Meaning that in the original example `ok google,` and `ok google, ` would both work.