Update using-audit-logs.md
"Generally speaking, people will actually want to use more than just the username of an audit log. It's best practice to actually return the whole entry and then take a property from it. that would make more sense." I've tried several things, setting the first entry to a variable using let, created an async function to return the first entry, no go. Any ideas as of how I could do this? It is a great idea.