New Coding Guide - Audit Logs (#13)
* Create AuditLogs.md

* Rename AuditLogs.md to understanding-and-using-audit-logs.md

* Update SUMMARY.md

* Update SUMMARY.md

* Update SUMMARY.md

* Update understanding-and-using-audit-logs.md

* Updated Spelling/Gramtical Errors

* Update understanding-and-using-audit-logs.md

* Update understanding-and-using-audit-logs.md

Adjusted spelling errors and corrected link(s)

* Update SUMMARY.md

* Rename understanding-and-using-audit-logs.md to using-audit-logs.md

* Update using-audit-logs.md

"Generally speaking, people will actually want to use more than just the username of an audit log. It's best practice to actually return the whole entry and then take a property from it. that would make more sense." I've tried several things, setting the first entry to a variable using let, created an async function to return the first entry, no go. Any ideas as of how I could do this? It is a great idea.

* Update using-audit-logs.md

Added a few checks to narrow down who deleted the message, whether it was the guy in audit logs or the author.

* Update using-audit-logs.md

* Update using-audit-logs.md

* Update SUMMARY.md

* Update SUMMARY.md

* Update SUMMARY.md

* Update SUMMARY.md

* Update SUMMARY.md

* Update SUMMARY.md