Add a helpful setInterval() example
This is helpful (especially for larger bots, but in any case) - it often happens that the game/activity doesn't update if there's certain Discord gateway reconnections or some outages. This just ensures that things are always updated. 

I apologize if it's out of place; I am happy to change any semantics as well.