Update heroku.md
the `web: echo 'words'` throws off heroku, makes it think it is a web process and did not work for me. just the `service:` did not work either. instead, the proposed change worked for me to deploy my discord bot.

also need to enable dynamo: https://stackoverflow.com/questions/51984638/discord-app-error-r10-when-deploying-with-heroku